Move FormConfigElementCard
