+++ packages/client/src/components/formConfig/Canvas.tsx
@@ -32,7 +32,7 @@ import { messages } from '@client/i18n/messages/views/formConfig'
 import { IStoreState } from '@client/store'
 import styled from '@client/styledComponents'
 import { useFieldDefinition } from '@client/views/SysAdmin/Config/Forms/hooks'
-import { FormConfigElementCard } from '@opencrvs/components/lib/interface/FormConfigElementCard'
+import { FormConfigElementCard } from '@opencrvs/components/lib/FormConfigElementCard'
 import React from 'react'
 import { useIntl } from 'react-intl'
 import { useDispatch, useSelector } from 'react-redux'
